Update Formatting Tools (#182)
* refactor(trove-classifiers): update to python 3.7-3.9

* refactor(linters): add `.flake8` and `.codespellrc` configuration files

Also add `PullRequest` to `ignore_words.txt`

* feat(pyproject): add `pyproject.toml`

Currently, this is only use for configuring linters and formatters (no
build file specifications are set). Configuration options for `black`
and `pydocstyle` are added here in this commit.

* feat(pre-commit): add pre-commit

* fix(mypy): add `mypy` dependency to `environment.yml`

This supports conda test/build environments.

* feat(pylint): Update `.pylintrc`

Make this match settings specififed in `Makefile`.

* feat(test): align checks

All checks are placed in appropriate config files. Test runner scripts
point to these so that there exists a single source of truth for all
configurations and that the tests performed across systems is the same.
This may be adjusted, of course, if different settings must be used on
a per-system basis.

* feat(requirements): add `toml`

This package supports reading `pyproject.toml`, which is required for
some of our tests.

* fix(pre-commit): fix file passing errors

`pre-commit` is known to override behavior from config files in
certain instances. This commit fixes known issues for:

- `isort`
- `black`
- `mypy`
- `pydocstyle`

For relevant details, see:

isort:
  + https://jugmac00.github.io/blog/isort-and-pre-commit-a-friendship-with-obstacles/
  + https://github.com/PyCQA/isort/issues/885

black:
  + https://github.com/psf/black/issues/1584

mypy/pydocstyle:
  + https://github.com/python/mypy/issues/4008#issuecomment-708060733
  + https://pre-commit.com/#hooks-pass_filenames

* feat(ignores): ignore specific linting errors

Ignore linting errors in source code in this PR. The purpose of this PR
is to update linting tools. A future PR will correct the errors. This is
done to separate concerns.

Co-authored-by: Hendry, Adam <adam.hendry@medtronic.com>